# swagger.yaml
openapi: 3.0.3
info:
  title: Cohortle API
  version: 1.0.0
  description: API documentation for Cohortle platform
servers:
  - url: https://cohortle-api.onrender.com/v1/api
    description: Production server
paths:
...
  /auth/register-email:
    post:
      summary: Register a new user with email
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: user@example.com
      responses:
        "200":
          description: Email sent successfully
        "400":
          description: Invalid input or email already used
        "500":
          description: Internal server error
  /auth/verify-email:
    post:
      summary: Verify user email using token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verify_token]
              properties:
                verify_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        "200":
          description: Email verified successfully
        "401":
          description: Invalid or expired token
        "500":
          description: Internal server error
  /auth/login:
    post:
      summary: Log in user with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: StrongPassword123
      responses:
        "200":
          description: Login successful
        "401":
          description: Invalid credentials
        "500":
          description: Internal server error
  /auth/forgot-password:
    post:
      summary: Request a password reset link
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: user@example.com
      responses:
        "200":
          description: Password reset link sent
        "401":
          description: Email not registered
        "500":
          description: Internal server error
  /auth/reset-password:
    post:
      summary: Reset password using token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  example: NewStrongPassword123
      responses:
        "200":
          description: Password reset successfully
        "401":
          description: Unauthorized or invalid token
        "500":
          description: Internal server error

  # ============================
  # Cohorts
  # ============================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...